Update my 'cleanbranches' alias.
Before the 'cleanbranches' alias was not properly excluding the master
branch. With this change, the master branch should be excluded from the
local delete whether I am on the master branch or not.

This also changes from the '-D' flag in 'git branch' to the '-d' flag.
The '-d' flag is safer because it does not delete the branch unless it
has already been merged into HEAD.